requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.2.0
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1
dj-database-url
psycopg2-binary
gunicorn
whitenoise